Move the CLI functiontionality to the CLI package
